+++ src/Value.js
@@ -1,4 +1,5 @@
 var React = require('react');
+var classes = require('classnames');
 
 var Value = React.createClass({
 
@@ -30,22 +31,33 @@ var Value = React.createClass({
 		}
 
 		if(!this.props.onRemove && !this.props.optionLabelClick) {
-			return <div className="Select-value">{label}</div>;
+			return (
+				<div
+					className={classes('Select-value', this.props.option.className)}
+					style={this.props.option.style}
+					title={this.props.option.title}
+				>{label}</div>
+			);
 		}
 
 		if (this.props.optionLabelClick) {
+
 			label = (
-				<a className="Select-item-label__a"
+				<a className={classes('Select-item-label__a', this.props.option.className)}
 					onMouseDown={this.blockEvent}
 					onTouchEnd={this.props.onOptionLabelClick}
-					onClick={this.props.onOptionLabelClick}>
+					onClick={this.props.onOptionLabelClick}
+					style={this.props.option.style}
+					title={this.props.option.title}>
 					{label}
 				</a>
 			);
 		}
 
 		return (
-			<div className="Select-item">
+			<div className={classes('Select-item', this.props.option.className)}
+				 style={this.props.option.style}
+				 title={this.props.option.title}>
 				<span className="Select-item-icon"
 					onMouseDown={this.blockEvent}
 					onClick={this.handleOnRemove}